01_pipeline_ml/pyproject.toml
[project]
name = "01-pipeline-ml"
version = "0.1.0"
description = "Pipeline de ML para churn prediction"
authors = [
    {name = "Miguel Mejia"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "pandas (>=2.3.1,<3.0.0)",
    "scikit-learn (>=1.7.1,<2.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "xgboost (>=3.0.3,<4.0.0)",
    "mlflow (>=3.2.0,<4.0.0)",
    "kfp (>=2.14.1,<3.0.0)",
    "google-cloud-aiplatform (>=1.107.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "ingestion", from = "src"}]

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.4.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
